Allows lockOptgroupOrder to be used with items not in a group
If some items had a property that put them to a group, but other items
did not have such a property, using lockOptgroupOrder caused a
ReferenceError.

After this commit, group order can be locked and items that do not belong
to any group have an order 0. Effectively this means that groups with
$order > 0 appear after non-grouped items, and groups with $order < 0 appear
before non-grouped items.